Split out flatten_cube tests
