+++ lib/iris/tests/unit/util/test_flatten_multidim_coord.py
@@ -26,8 +26,7 @@ import iris.tests as tests
 import unittest
 
 import iris.tests.stock as stock
-from iris.coords import DimCoord, AuxCoord
-from iris.util import flatten_multidim_coord, flatten_cube
+from iris.util import flatten_multidim_coord
 
 
 class Test(tests.IrisTest):
@@ -47,48 +46,6 @@ class Test(tests.IrisTest):
         cube_b = flatten_multidim_coord(cube_a, coord)
         self.assertEqual(cube_b.dim_coords, (cube_a.coord('wibble'), ))
         self.assertEqual(cube_b.shape, (2, 12))
-
-    def test_simple_cube_flattened(self):
-        cube_a = stock.simple_2d()
-        cube_b = flatten_cube(cube_a)
-        self.assertEqual(cube_b.dim_coords, tuple())
-        self.assertEqual(cube_b.shape, (12, ))
-
-    def test_multiple_oned_dim_coords_flattened(self):
-        cube_a = stock.simple_3d_w_multidim_coords()
-        cube_a.add_dim_coord(DimCoord([0, 1, 2], var_name='blah'), (1,))
-        cube_a.add_dim_coord(DimCoord([0, 1, 2, 3], var_name='blah2'), (2,))
-        cube_b = flatten_cube(cube_a, (1, 2))
-        self.assertEqual(cube_b.dim_coords, (cube_a.coord('wibble'), ))
-        self.assertEqual(cube_b.coord('blah', dim_coords=False).shape, (12, ))
-        self.assertEqual(cube_b.coord('blah2', dim_coords=False).shape, (12,))
-        self.assertEqual(cube_b.shape, (2, 12))
-
-    def test_oned_aux_coord_flattened(self):
-        cube_a = stock.simple_3d_w_multidim_coords()
-        cube_a.add_aux_coord(AuxCoord([0, 1, 2], var_name='blah'), (1,))
-        cube_b = flatten_cube(cube_a, (1, 2))
-        self.assertEqual(cube_b.dim_coords, (cube_a.coord('wibble'), ))
-        self.assertEqual(cube_b.coord('blah', dim_coords=False).shape, (12, ))
-        self.assertEqual(cube_b.shape, (2, 12))
-
-    def test_aux_coords_leading(self):
-        # Check that it doensn't matter which dimensions the aux-coord spans
-        cube_a = stock.simple_3d_w_multidim_coords()
-        # Move the aux coord dims to the front of the cube
-        cube_a.transpose((1, 2, 0))
-        cube_b = flatten_cube(cube_a, (0, 1))
-        self.assertEqual(cube_b.dim_coords, (cube_a.coord('wibble'), ))
-        self.assertEqual(cube_b.shape, (12, 2))
-
-    def test_split_aux_coord_dims(self):
-        # Check that an aux-coord with 'split' (non-continuous) dimensions are
-        #  flattened correctly
-        cube_a = stock.simple_3d_w_multidim_coords()
-        cube_a.transpose((1, 0, 2))
-        cube_b = flatten_cube(cube_a, (0, 2))
-        self.assertEqual(cube_b.dim_coords, (cube_a.coord('wibble'), ))
-        self.assertEqual(cube_b.shape, (12, 2))
 
 
 if __name__ == '__main__':
